requirements.txt
python-parallel-collections>=1.2.1
